Moved to original script
removed enemy controller script copy and made changes to original script
