--- Assets/Scripts/Enemy/FireGoombaControllerUpdated.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a4d58705b1c2c944698c94d3aac205fb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
